requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyarrow>=12.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0
optuna>=3.0.0
flask>=3.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
sqlalchemy>=2.0.0


